Finish project

+++ src/components/login/Login.js
@@ -11,13 +11,14 @@ import { Formik } from 'formik'
 import * as Yup from 'yup'
 import axios from 'axios'
 import { useDispatch, useSelector } from 'react-redux'
-import { setToken, setUsername, deletePassword } from '../../redux/slices/accountSlice'
+import { setUsername, deletePassword, setRole, setUserId } from '../../redux/slices/accountSlice'
 import { BASE_URL_USER } from '../../constants/api'
 import FORMTextInput from '../_fragments/FORMTextInput'
 import FORMPasswordInput from '../_fragments/FORMPasswordInput'
 import SubmitButton from '../_fragments/FORMSubmitButton'
 import SocialLoginComponent from './SocialLogin'
 import { ROLE_ADMIN, ROLE_HOST } from '../../constants/roles'
+import { loginSetup, logoutAPI } from '../../services/authService'
 
 export default function Login() {
   const navigate = useNavigate()
@@ -39,7 +40,7 @@ export default function Login() {
     if (localStorage.getItem('loggedIn')) {
       navigate('/')
     }
-  }, [])
+  }, [navigate])
 
   const validationSchema = Yup.object().shape({
     username: Yup.string()
@@ -57,18 +58,20 @@ export default function Login() {
       username: values.username,
       password: values.password,
     })
-      .then((res) => {
+      .then(async (res) => {
         const user = res.data;
+        localStorage.setItem("token", user.token);
+        if (user.userStatus === 'LOCKED') {
+          await logoutAPI()
+          toast.warning("Your account has been LOCKED, please contact Admin.", { hideProgressBar: true })
+          return navigate('/login');
+        }
         const role = user.authorities[0].authority;
-        dispatch(setToken(user.token))
         dispatch(setUsername(user.username))
+        dispatch(setUserId(user.id))
+        dispatch(setRole(role))
         dispatch(deletePassword())
-        localStorage.setItem('userId', user.id)
-        localStorage.setItem('token', user.token)
-        localStorage.setItem('loggedIn', JSON.stringify(true))
-        localStorage.setItem('username', user.username)
-        localStorage.setItem('role', role)
-        toast.success('login successful', { hideProgressBar: true })
+        await loginSetup(user, role)
         if (role === ROLE_ADMIN) {
           return navigate('/admin');
         }
@@ -112,16 +115,20 @@ export default function Login() {
                       <SubmitButton
                         label="Login"
                       />
-                      <CRow className="mb-3 mt-3">
+                      <CRow className="mb-1 mt-3">
                         <CCol className="d-flex justify-content-center fs-6">
                           Don't have an account?
-                          <Link to="/register" style={{textDecoration: "none"}}> Register</Link>
+                          <Link to="/register" style={{ textDecoration: 'none' }}>Register</Link>
                         </CCol>
                       </CRow>
-                      <hr/>
-                      <CRow>
+                      <div className="d-flex align-items-center">
+                        <hr className="flex-grow-1" />
+                        <span className="mx-2 text-muted">Or</span>
+                        <hr className="flex-grow-1" />
+                      </div>
+                      <CRow className="mt-2">
                         <CCol className="d-flex justify-content-center">
-                          <SocialLoginComponent/>
+                          <SocialLoginComponent />
                         </CCol>
                       </CRow>
                     </CForm>
